create whls for deps for pyright/verifytypes
