+++ eng/tox/tox.ini
@@ -129,9 +129,7 @@ setenv =
   {[testenv]setenv}
   PROXY_URL=http://localhost:5018
 deps =
-  -r ../../../eng/test_tools.txt
-  -e ../../../tools/azure-sdk-tools
-  -e ../../../tools/azure-devtools
+  {[base]deps}
   pyright==1.1.287
 commands =
     {envbindir}/python {toxinidir}/../../../eng/tox/create_package_and_install.py \
@@ -139,6 +137,8 @@ commands =
       -p {toxinidir} \
       -w {envtmpdir} \
       --package-type sdist
+    {envbindir}/python {toxinidir}/../../../eng/tox/create_whl_for_dependencies.py \
+      -p {toxinidir}
     {envbindir}/python {toxinidir}/../../../eng/tox/run_pyright.py -t {toxinidir}
 
 [testenv:verifytypes]
@@ -150,9 +150,7 @@ setenv =
   {[testenv]setenv}
   PROXY_URL=http://localhost:5019
 deps =
-  -r ../../../eng/test_tools.txt
-  -e ../../../tools/azure-sdk-tools
-  -e ../../../tools/azure-devtools
+  {[base]deps}
   pyright==1.1.287
 commands =
     {envbindir}/python {toxinidir}/../../../eng/tox/create_package_and_install.py \
@@ -160,6 +158,8 @@ commands =
       -p {toxinidir} \
       -w {envtmpdir} \
       --package-type sdist
+    {envbindir}/python {toxinidir}/../../../eng/tox/create_whl_for_dependencies.py \
+      -p {toxinidir}
     {envbindir}/python {toxinidir}/../../../eng/tox/run_verifytypes.py -t {toxinidir}
